src/requirements/base.txt
alembic==1.13.1
astroid==3.3.4
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored==2.2.4
coverage==7.6.1
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
et-xmlfile==1.1.0
fabric==3.2.2
Faker==28.1.0
fire==0.7.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.10
iniconfig==2.0.0
invoke==2.2.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
paramiko==3.4.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pylint==3.3.1
PyNaCl==1.5.0
pytest==8.2.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
rich==13.9.4
six==1.16.0
termcolor==2.5.0
tomlkit==0.13.2
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.3
wrapt==1.16.0
